requirements.txt
PyYAML
discord.py
requests
pillow
aiohttp[speedups]
pydantic
sqlalchemy
quart
